fix: CI — replace x_forwarded_for enum with string literal, fix narrowing cast
- boost::beast::http::field::x_forwarded_for not available in older Boost
- Use raw X-Forwarded-For header string instead
- Cast util::now_ms() uint64_t to int64_t for narrowing compatibility

+++ src/progressive/ratelimit/ratelimit.cpp
@@ -26,7 +26,7 @@ bool RateLimiter::allow(std::string_view key) {
 
   auto it = buckets_.find(key);
   if (it == buckets_.end()) {
-    buckets_[std::string(key)] = {burst_ - 1, util::now_ms()};
+    buckets_[std::string(key)] = {burst_ - 1, static_cast<int64_t>(util::now_ms())};
     return true;
   }
